[NT-0] fix: Change underlying type of `ComponentType` to something representable as a JavaScript number (#940)
The changes added in 61f82909b46e5c504444d90c2f4ea0a3c2687bb0 made
`ComponentType` explicitly a `uint64_t`, to be more explicit about
the existing wire constraints that this ID ultimately has.

This wasn't intended to have any consequences for existing consumers,
but on web platforms, this forces a BigInt in the JS interop code that
Emscripten generates (which is actually more honest), which don't
strictly compare equal to the number types in the generated enum
i.e. expressions like: `c.getComponentType() === ComponentType.Audio`,
or when used in a switch statement. So, this easily breaks existing code.

For now, mitigate this by using a smaller type (`uint16_t` which is big
enough for our current range, supported by the wrapper gen already, and
explicitly unsigned vs the default `int` underlying type, which is the
main point here).

I still think being explicit about these constraints is a good thing,
but if others disagree I can just get rid of all of the static checking
for this and put it back to how it was before.

+++ Library/include/CSP/Multiplayer/ComponentBase.h
@@ -52,7 +52,7 @@ class ComponentScriptInterface;
 /// @brief Represents the type of component.
 ///
 /// Values with _DEPRECATED appended to their name should not be used. They are retained only for backwards compatibility.
-enum class ComponentType : uint64_t
+enum class ComponentType : uint16_t
 {
     Invalid = 0,
     Core = 1,
